requirement.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
Naked==0.1.31
numpy==1.14.0
pandas==0.22.0
pycrypto==2.6.1
PyMySQL==0.8.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
six==1.11.0
urllib3==1.22
